Allow for newlines to have a hanging indent if the next non-empty line has a hanging indent.
This is necessary for a nested keep-sorted test case that uses group=yes
to capture the nested keep-sorted blocks as well as newline_separated=yes
in one of those nested blocks.

Without this, those newlines that the nested block introduces break up
the hanging indent grouping, resulting in pretty bad output.