refactor: Move sorting logic from block.go to line_group.go.
This also moves the logic a bit more towards an imperative style from
the current functional style, which might be a bit easier to reason
about moving forward.

Also note that I'm caching calculated values and keeping track of which
values are calculated. Caching the calculated values should perhaps
improve performance (not that we've had any user complaints about
speed). I intend to use the fact that we're keeping track of which
values are calculated to improve the debuggability of keep-sorted in a
follow-up.