refactor: Make keep-sorted debug output a bit more useful for figuring out why things were sorted a certain way. (#88)
The debug output now reports all of the tokens that were actually used
in comparisons. It also now reports all of the line groups in
their sorted order instead of the line groups in their original order,
but I think that should be fine.

If you want to see it in action, run something like

```sh
$ go run . -vv --id keep-sorted-test - < <(cat goldens/*.in) 1>/dev/null
```