Adding user-defined group delimiters
This addresses https://github.com/google/keep-sorted/issues/97. The
discussion there suggested trying to override line breaks, but this
turns out to be difficult. The options are not specified until after
we've split into lines, and lots of places use line numbers for fixes.

Instead, we allow customization of how lines are associated into
groups. In order to support the TOML's use case (end groups on blank
line) and SQL's use case (end groups if the line ends in a semicolon),
we allow a new list of regexes, group_delimiter_regexes that determine
if the line should end a group.

There's a subtlety to precedence with sticky_prefixes. Since lines
starting with sticky_prefixes should associate to a group below them,
the first line with a sticky prefix can create a new group, despite what
the We allow sticky_prefixes to take precedence.