refactor: Simplify some conditional logic to remove unreachable branches. (#105)
```
if !lineRange.empty() {
  finishGroup()
}
```

guarantees that `lineRange.empty()` will be true after it (`finishGroup`
zeroes out `lineRange`). That lets us get rid of one of the branches in
the `metadata.opts.Group` conditional. But now the remaining branch
shares a lot in common with the `else` branch, so we can simplify a bit
further.